review workflow: move library modules to lib/ part 2
