+++ packages/workflows/src/review/lib/reviewWorkflow.ts
@@ -3,12 +3,12 @@
 import { CoreEvents, EventBus } from "@prsense/core";
 import type { DiffProvider } from "@prsense/core";
 import type { ResolvedConfig, CredentialContext } from "@prsense/config";
-import type { ReviewWorkflowResult } from "./types.js";
-import { loadDiff } from "./steps/loadDiff.js";
-import { resolveContext } from "./steps/resolveContext.js";
-import { createLlmClientSafe } from "./steps/createLlmClient.js";
-import { runReview } from "./steps/runReview.js";
-import { finalizeSignals } from "./steps/finaliseSignals.js";
+import type { ReviewWorkflowResult } from "../types.js";
+import { loadDiff } from "../steps/loadDiff.js";
+import { resolveContext } from "../steps/resolveContext.js";
+import { createLlmClientSafe } from "../steps/createLlmClient.js";
+import { runReview } from "../steps/runReview.js";
+import { finalizeSignals } from "../steps/finaliseSignals.js";
 import { IndexMetadataRepository, RagChunkRepository } from "@prsense/context";
 
 export async function runReviewWorkflow({
