+++ packages/workflows/src/review/lib/runFileReview.ts
@@ -3,9 +3,9 @@
 import { CoreEvents, DiffFile, type EventBus } from "@prsense/core";
 import { ResolvedConfig } from "@prsense/config";
 import { buildReviewPrompt } from "@prsense/core";
-import { extractJson } from "../extractJson.js";
+import { extractJson } from "./extractJson.js";
 import type { FileReviewResult, ReviewMetadata } from "../types.js";
-import { validateReviewOutput } from "../validateReviewOutput.js";
+import { validateReviewOutput } from "./validateReviewOutput.js";
 import { LlmClient } from "@prsense/llm";
 
 export async function runFileReview({
